backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
Django==4.2.13
djangorestframework==3.15.2
django-cors-headers==4.3.1

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# ─── Twilio Integration ───────────────────────────────────────────────────────
twilio==9.0.4

# ─── Environment Variables ────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Development / Utilities ──────────────────────────────────────────────────
gunicorn==22.0.0
